+++ e2e/app.e2e-spec.ts
@@ -1,7 +1,7 @@
 
 describe('QuickStart E2E Tests', function () {
 
-  let expectedMsg = 'My First Angular 2 App';
+  let expectedMsg = 'My First Angular App';
 
 
   beforeEach(function () {